@@ -5,21 +5,20 @@ import java.net.URISyntaxException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import javax.inject.Inject;
-import javax.json.bind.JsonbBuilder;
-import javax.json.bind.JsonbConfig;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response.Status;
-import javax.ws.rs.core.UriBuilder;
-import javax.ws.rs.ext.Providers;
+import jakarta.inject.Inject;
+import jakarta.json.bind.JsonbBuilder;
+import jakarta.json.bind.JsonbConfig;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response.Status;
+import jakarta.ws.rs.core.UriBuilder;
+import jakarta.ws.rs.ext.Providers;
 
 import org.assertj.core.api.Assertions;
 import org.jboss.resteasy.client.jaxrs.ResteasyWebTarget;
-import org.jboss.resteasy.core.Dispatcher;
 import org.jboss.resteasy.mock.MockDispatcherFactory;
 import org.jboss.resteasy.mock.MockHttpRequest;
 import org.jboss.resteasy.mock.MockHttpResponse;
-import org.jboss.resteasy.plugins.providers.jsonb.JsonBindingProvider;
+import org.jboss.resteasy.spi.Dispatcher;
 import org.jboss.weld.junit4.WeldInitiator;
 import org.junit.Before;
 import org.junit.BeforeClass;
@@ -64,7 +63,7 @@ public class WeldSampleTest extends RestClientTestBase {
 	private Dispatcher dispatcher;
 	
 	
-	JsonBindingProvider provider;
+	// JsonBindingProvider provider;
 	
 	@Before
 	public void init() {
@@ -81,7 +80,7 @@ public class WeldSampleTest extends RestClientTestBase {
 		req.setSampleModel(sampleModel);
 
 		dispatcher.getProviderFactory().getProviderInstances().forEach(o -> System.out.println(o.getClass().toString()));
-		dispatcher.getProviderFactory().getProviderInstances().stream().filter(e -> e instanceof JsonBindingProvider).forEach(e -> System.out.println(e.toString()));
+		// dispatcher.getProviderFactory().getProviderInstances().stream().filter(e -> e instanceof JsonBindingProvider).forEach(e -> System.out.println(e.toString()));
 		//dispatcher.getProviderFactory().getProviderClasses().forEach(e -> System.out.println(e.toString()));
 		//ContextResolver<Jsonb> contextResolver = providers.getContextResolver(Jsonb.class, MediaType.APPLICATION_JSON_TYPE);
 		
