@@ -1,11 +1,11 @@
 package com.example.app.rest.model;
 
-import javax.validation.constraints.NotNull;
-
 import com.example.app.com.json.bind.annotation.MaskedString;
 
+import jakarta.validation.constraints.NotNull;
+
 public class SampleModel {
-	
+
 	@NotNull(message = "{APYC00001}")
 	@MaskedString
 	private String gooooal;
