@@ -1,12 +1,12 @@
 package com.example.app.rest.model;
 
-import javax.json.bind.annotation.JsonbTransient;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-
 import org.apache.commons.lang3.StringUtils;
 
 import com.example.app.com.json.bind.annotation.MaskedString;
+
+import jakarta.json.bind.annotation.JsonbTransient;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 
 public class SampleRequest {
 
@@ -16,14 +16,14 @@ public class SampleRequest {
 	// @JsonbTypeSerializer(CustomJsonbSerializer.class)
 	@MaskedString
 	private String businessData;
-	
+
 	@NotNull(message = "{APYC00001}")
 	private String correlationData;
-	
+
 	@NotNull(message = "{APYC00001}")
 	@Valid
 	private SampleModel sampleModel;
-	
+
 	// @AssertTrue(message = "{APYC00004}")
 	@JsonbTransient
 	public boolean isCorrelationCheck() {
